requirements.txt
flask~=3.1.2
flask-login~=0.6.3
werkzeug~=3.1.3
python-dotenv~=1.2.1